requirements.txt
Flask
fpdf
pdfplumber
python-dotenv
google-generativeai
groq
requests
Werkzeug
google-genai